requirements.txt
pandas
requests
beautifulsoup4
transformers
bertopic
datasketch
nltk
spacy
vaderSentiment
scipy
newspaper3k
lxml_html_clean
torch
sentence-transformers
